* Change default rings gap + new customizer preset
I find that a 0.6mm gap between the rings works much better on my
printer, so that is the new default. I think the original uses 0.5mm,
but that is a bit too tight for my current setup.

Also add a new OpenSCAD customizer profile - I love the new ability to
add custom presets to Customizer! :-)