BitHolder: New features and Combine OpenSCAD and Thingiverse versions into one.
New features:
 * Gaps between holes can be adjusted
 * Edge widths can be adjusted
 * Hole sizes can be set
 * Holes can be round or hexagonal

Due to the fillets.scad dependancy, and Thingiverse not allowing
importing external libs, there used to be the main version that imported
the fillets.scad lib, and the Thingiverse version that had the fillets
functions embedded.
This was tedious to maintain, so fillet functions are now embedded and
the other version has been deleted. Attribution to the fillets lib is
given in the source.

Also removed the scaling option for printing.